app/requirements.txt
python-dotenv==1.1.0
openai==1.82.0
PyPDF2==3.0.1
langchain_text_splitters==0.3.8
langchain-openai==0.3.18
scikit-learn==1.6.1
